# oapispec/swagger.yaml
basePath: /
definitions:
...
  admin.AdminUpdateOrgRequest:
    properties:
      description:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
    required:
    - name
    - settings
    type: object
  admin.AdminUpdateOrgResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  config.APIKeyConfig:
    properties:
      description:
        description: Description of the key
        type: string
      name:
        description: The name of the API Key
        type: string
    required:
    - description
    - name
    type: object
  config.AuditLogsConfig:
    properties:
      enable:
        description: Enable or disable audit logs.
        type: boolean
    type: object
  config.BrandingConfig:
    properties:
      appName:
        description: AppName is the name of the application, used in various places
          like email templates, UI, etc.
        type: string
      companyName:
        description: CompanyName is the name of the company, used in emails, UI, etc.
        type: string
      companyNameShort:
        description: CompanyNameShort is a short version of the company name, used
          in places where space is limited.
        type: string
      supportURL:
        description: SupportURL is the URL for support, used in emails, UI, etc.
        type: string
    required:
    - appName
    - companyName
    - companyNameShort
    - supportURL
    type: object
  config.CompleteConfig:
    properties:
      branding:
        $ref: '#/definitions/config.BrandingConfig'
      debug:
        description: Debug mode for the application
        type: boolean
      jwt:
        $ref: '#/definitions/config.JWTConfig'
      multitenancy:
        type: boolean
      notifications:
        $ref: '#/definitions/config.NotificationsConfig'
      public:
        $ref: '#/definitions/config.PublicConfig'
      security:
        $ref: '#/definitions/config.SecurityConfig'
    required:
    - branding
    - jwt
    - multitenancy
    - notifications
    - public
    - security
    type: object
  config.EmailEndpointsConfig:
    properties:
      passwordReset:
        description: |-
          PasswordReset is the endpoint for the password reset link.
          A `token` parameter will be passed to this URL.
          Pass a full url, eg. https://auth.example.com/password-reset
        type: string
      verificationEmail:
        description: |-
          VerificationEmail is the endpoint for the email verification link.
          A `token` parameter will be passed to this URL.
          Pass a full url, eg. https://auth.example.com/verify-email
        type: string
    required:
    - passwordReset
    - verificationEmail
    type: object
  config.EmailNotificationConfig:
    properties:
      endpoints:
        allOf:
        - $ref: '#/definitions/config.EmailEndpointsConfig'
        description: Endpoints holds the configuration for URLs inside emails.
      provider:
        description: Provider is the email provider to use for sending emails.
        enum:
        - ses
        - sendgrid
        - smtp
        type: string
    required:
    - endpoints
    - provider
    type: object
  config.JWTConfig:
    properties:
      refreshTokenExpiration:
        description: 'Refresh token expiration time in seconds (default: 30d, 2592000)'
        minimum: 86400
        type: integer
      sessionTokenExpiration:
        description: 'Session token expiration time in seconds (default: 1hr, 3600)'
        minimum: 60
        type: integer
    required:
    - refreshTokenExpiration
    - sessionTokenExpiration
    type: object
  config.NotificationsConfig:
    properties:
      email:
        allOf:
        - $ref: '#/definitions/config.EmailNotificationConfig'
        description: Email configuration for sending notifications
      sms:
        allOf:
        - $ref: '#/definitions/config.SMSNotificationConfig'
        description: SMS Configuration for sending notifications
    required:
    - email
    type: object
  config.PublicConfig:
    properties:
      debugBaseURL:
        type: string
      domain:
        description: |-
          Domain is the domain at which Nexeres UI is being hosted.

          Eg. if hosting Nexeres at auth.example.com, provide "example.com" here.

          This setting is used as the Domain for all Cookies (except Refresh Token Cookie, that is only set at "subdomain.domain")
        type: string
      scheme:
        description: |-
          Scheme for public URLs (http/https)

          Use "https" even if TLS termination is handled by reverse proxy
        enum:
        - http
        - https
        type: string
      subDomain:
        description: |-
          Subdomain at which Nexeres UI is being hosted.

          Eg. If hosting Nexeres at auth.example.com, provide "auth" here.
        type: string
    required:
    - domain
    - scheme
    - subDomain
    type: object
  config.RateLimitConfig:
    properties:
      rate:
        description: |-
          Rate limit for API requests.
          Format: "R-U", where R is requests and U is the time unit (s - per second, m - per minute, h - per hour, d - per day)
        type: string
    required:
    - rate
    type: object
  config.SMSNotificationConfig:
    properties:
      provider:
        description: 'TODO: Implement SMSNotificationConfig'
        enum:
        - twilio
        type: string
    required:
    - provider
    type: object
  config.SecurityConfig:
    properties:
      apiKeys:
        description: |-
          The list of API Keys which are allowed to access the API endpoints.
          Requests without an API key, or with a key not specified here, will be denied with 401.
        items:
          $ref: '#/definitions/config.APIKeyConfig'
        type: array
      auditLogs:
        allOf:
        - $ref: '#/definitions/config.AuditLogsConfig'
        description: Enable or disable audit logs.
      rateLimit:
        allOf:
        - $ref: '#/definitions/config.RateLimitConfig'
        description: Rate limiting configuration.
    required:
    - apiKeys
    - auditLogs
    - rateLimit
    type: object
  db.GetMinimalRolesForOrgRow:
    properties:
      createdAt:
        type: string
      id:
        type: string
      roleDesc:
        type: string
      roleName:
        type: string
      updatedAt:
        type: string
    type: object
  db.Org:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  db.OrgDomain:
    properties:
      autoJoin:
        type: boolean
      autoJoinRoleId:
        type: string
      autoJoinRoleName:
        type: string
      createdAt:
        type: string
      domain:
        type: string
      orgId:
        type: string
      updatedAt:
        type: string
      verified:
        type: boolean
      verifiedAt:
        type: string
    type: object
...
  models.OrgMFASettings:
    properties:
      required:
        type: boolean
      roles:
        additionalProperties:
          type: string
        description: role name to role-id mapping
        type: object
    type: object
  models.OrgSettings:
    properties:
      mfa:
        $ref: '#/definitions/models.OrgMFASettings'
    type: object
  reserr.ErrorResponse:
    properties:
      debug:
        description: DebugMessage is a technical message that can be used for debugging.
        type: string
      message:
        description: Message is a user-friendly message that can be displayed to the
          end user.
        type: string
    type: object
  resquery.FilterMode:
    enum:
    - AND
    - OR
    type: string
    x-enum-varnames:
    - FilterModeAND
    - FilterModeOR
  resquery.FilterOption:
    properties:
      field:
        type: string
      op:
        $ref: '#/definitions/resquery.QueryFilterOp'
      value: {}
    type: object
  resquery.QueryFilterOp:
    enum:
    - contains
    - equals
    - lte
    - gte
    - lt
    - gt
    type: string
    x-enum-varnames:
    - OpContains
    - OpEquals
    - OpLte
    - OpGte
    - OpLt
    - OpGt
  resquery.QueryFilters:
    properties:
      mode:
        $ref: '#/definitions/resquery.FilterMode'
      options:
        items:
          $ref: '#/definitions/resquery.FilterOption'
        minItems: 1
        type: array
    required:
    - mode
    - options
    type: object
  resquery.QueryPagination:
    properties:
      page:
        description: Page is the page number to retrieve (0-based).
        minimum: 0
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - page
    - pageSize
    type: object
  resquery.SortOption:
    properties:
      direction:
        enum:
        - asc
        - desc
        - ASC
        - DESC
        type: string
      field:
        type: string
    required:
    - direction
    - field
    type: object
  sys_admin.SysAdminCreateOrgRequest:
    properties:
      avatarURL:
        type: string
      description:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
      slug:
        type: string
    required:
    - name
    - settings
    - slug
    type: object
  sys_admin.SysAdminCreateOrgResponse:
    properties:
      message:
        type: string
      orgId:
        type: string
      success:
        type: boolean
    type: object
  sys_admin.SysAdminDeleteOrgResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  sys_admin.SysAdminGetConfigResponse:
    properties:
      config:
        $ref: '#/definitions/config.CompleteConfig'
      message:
        type: string
      success:
        type: boolean
    type: object
  sys_admin.SysAdminGetOrgDetailsResponse:
    properties:
      domains:
        items:
          $ref: '#/definitions/db.OrgDomain'
        type: array
      message:
        type: string
      org:
        $ref: '#/definitions/db.Org'
      roles:
        items:
          $ref: '#/definitions/db.GetMinimalRolesForOrgRow'
        type: array
      success:
        type: boolean
    type: object
  sys_admin.SysAdminListOrgsRequest:
    properties:
      filters:
        $ref: '#/definitions/resquery.QueryFilters'
      pagination:
        $ref: '#/definitions/resquery.QueryPagination'
      sort:
        items:
          $ref: '#/definitions/resquery.SortOption'
        type: array
    required:
    - sort
    type: object
  sys_admin.SysAdminListOrgsResponse:
    properties:
      message:
        type: string
      orgs:
        description: List of organizations
        items:
          $ref: '#/definitions/db.Org'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/resquery.QueryPagination'
        description: Pagination details, default if not provided in the request
      success:
        type: boolean
      total:
        description: Total number of organizations matching the filters (if any)
        type: integer
    type: object
...
host: localhost:3360
info:
  contact:
    email: contact@nbrglm.com
    name: NBRGLM Support
    url: https://nbrglm.com/support
  description: This is the NBRGLM Nexeres API Spec documentation.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://nbrglm.com/nexeres/terms
  title: NBRGLM Nexeres API Spec
  version: 0.0.1
paths:
...
  /api/admin/orgs/{orgId}:
    post:
      consumes:
      - application/json
      description: Handles AdminUpdateOrg requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUpdateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminUpdateOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminUpdateOrg Endpoint
      tags:
      - admin
...
  /api/sys/admin/config:
    get:
      consumes:
      - application/json
      description: Handles SysAdminGetConfig requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminGetConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminGetConfig Endpoint
      tags:
      - sys_admin
  /api/sys/admin/orgs:
    post:
      consumes:
      - application/json
      description: Handles SysAdminListOrgs requests
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys_admin.SysAdminListOrgsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminListOrgsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminListOrgs Endpoint
      tags:
      - sys_admin
    put:
      consumes:
      - application/json
      description: Handles SysAdminCreateOrg requests
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys_admin.SysAdminCreateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminCreateOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminCreateOrg Endpoint
      tags:
      - sys_admin
  /api/sys/admin/orgs/{orgId}:
    delete:
      consumes:
      - application/json
      description: Handles SysAdminDeleteOrg requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminDeleteOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminDeleteOrg Endpoint
      tags:
      - sys_admin
    get:
      consumes:
      - application/json
      description: Handles SysAdminGetOrgDetails requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminGetOrgDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminGetOrgDetails Endpoint
      tags:
      - sys_admin
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-NEXERES-API-Key
    type: apiKey
  AdminHeaderAuth:
    in: header
    name: X-NEXERES-Admin-Token
    type: apiKey
  RefreshHeaderAuth:
    in: header
    name: X-NEXERES-Refresh-Token
    type: apiKey
  SessionHeaderAuth:
    in: header
    name: X-NEXERES-Session-Token
    type: apiKey
swagger: "2.0"